2019.6.12

+++ variables/variables_functions.cpp
@@ -4,7 +4,7 @@
 
 namespace variables {
 	void init_variables_state(world_state& ws) {
-		if(ws.w.variable_s.global_variables.size() != ws.s.variables_m.count_global_variables)
+		if(ws.w.variable_s.global_variables.size() != int32_t(ws.s.variables_m.count_global_variables))
 			ws.w.variable_s.global_variables.resize(ws.s.variables_m.count_global_variables);
 	}
 	void reset_state(variables_state& s) {
